Rename files/actions to be more descriptive
