+++ project.yaml
@@ -5,6 +5,24 @@ expectations:
   population_size: 1000
 
 actions:
+
+  generate_dataset_definition_table1:
+    run: >
+      ehrql:v1
+       generate-dataset analysis/dataset_definition_table1.py
+       --test-data-file analysis/test_dataset_definition_table1.py
+       --output output/population/pf_table1.csv.gz
+    outputs:
+      highly_sensitive:
+        cohort: output/population/pf_table1.csv.gz
+
+  create_table1:
+    run: r:latest analysis/create_tables.R
+    needs: [generate_dataset_definition_table1]
+    outputs:
+      moderately_sensitive:
+        dataset: output/population/pf_demographics.csv
+
   generate_pf_breakdown_measures:
     run: >
       ehrql:v1 generate-measures analysis/measures_definition_pf_breakdown.py
@@ -39,16 +57,6 @@ actions:
       moderately_sensitive:
         measure: output/measures/pf_descriptive_stats_measures.csv
 
-  generate_pf_dataset_definition:
-    run: >
-      ehrql:v1
-       generate-dataset analysis/pf_dataset_definition.py
-       --test-data-file analysis/test_pf_dataset_definition.py
-       --output output/population/pf_population.csv.gz
-    outputs:
-      highly_sensitive:
-        cohort: output/population/pf_population.csv.gz
-
   generate_pf_med_counts_measures:
     run: >
       ehrql:v1 generate-measures analysis/measures_definition_pf_consultation_med_counts.py
@@ -57,10 +65,3 @@ actions:
     outputs:
       moderately_sensitive:
         measure: output/measures/consultation_med_counts_measures.csv
-
-  pf_demographics_table:
-    run: r:latest analysis/dataset_definition_tables.R
-    needs: [generate_pf_dataset_definition]
-    outputs:
-      moderately_sensitive:
-        dataset: output/population/pf_demographics.csv
